Sourcery refactored main branch

+++ tg_bot/modules/cleaner.py
@@ -10,11 +10,7 @@ from tg_bot.modules.disable import DisableAbleCommandHandler
 from tg_bot.modules.helper_funcs.chat_status import user_admin, bot_can_delete, dev_plus, connection_status
 from tg_bot.modules.sql import cleaner_sql as sql
 
-if ALLOW_EXCL:
-    CMD_STARTERS = ('/', '!')
-else:
-    CMD_STARTERS = ('/')
-
+CMD_STARTERS = ('/', '!') if ALLOW_EXCL else '/'
 BLUE_TEXT_CLEAN_GROUP = 15
 CommandHandlerList = (CommandHandler, CustomCommandHandler, DisableAbleCommandHandler)
 command_list = ["cleanblue", "ignoreblue", "unignoreblue", "listblue", "ungignoreblue", "gignoreblue"
@@ -32,21 +28,20 @@ def clean_blue_text_must_click(bot: Bot, update: Update):
     chat = update.effective_chat
     message = update.effective_message
 
-    if chat.get_member(bot.id).can_delete_messages:
-        if sql.is_enabled(chat.id):
-            fst_word = message.text.strip().split(None, 1)[0]
+    if chat.get_member(bot.id).can_delete_messages and sql.is_enabled(chat.id):
+        fst_word = message.text.strip().split(None, 1)[0]
 
-            if len(fst_word) > 1 and any(fst_word.startswith(start) for start in CMD_STARTERS):
+        if len(fst_word) > 1 and any(fst_word.startswith(start) for start in CMD_STARTERS):
 
-                command = fst_word[1:].split('@')
-                chat = update.effective_chat
+            command = fst_word[1:].split('@')
+            chat = update.effective_chat
 
-                ignored = sql.is_command_ignored(chat.id, command[0])
-                if ignored:
-                    return
+            ignored = sql.is_command_ignored(chat.id, command[0])
+            if ignored:
+                return
 
-                if command[0] not in command_list:
-                    message.delete()
+            if command[0] not in command_list:
+                message.delete()
 
 
 @run_async
@@ -58,14 +53,14 @@ def set_blue_text_must_click(bot: Bot, update: Update, args: List[str]):
     chat = update.effective_chat
     message = update.effective_message
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
-        if val == "off" or val == "no":
+        if val in ["off", "no"]:
             sql.set_cleanbt(chat.id, False)
             reply = "Bluetext cleaning has been disabled for <b>{}</b>".format(html.escape(chat.title))
             message.reply_text(reply, parse_mode=ParseMode.HTML)
 
-        elif val == "yes" or val == "on":
+        elif val in ["yes", "on"]:
             sql.set_cleanbt(chat.id, True)
             reply = "Bluetext cleaning has been enabled for <b>{}</b>".format(html.escape(chat.title))
             message.reply_text(reply, parse_mode=ParseMode.HTML)
@@ -75,10 +70,7 @@ def set_blue_text_must_click(bot: Bot, update: Update, args: List[str]):
             message.reply_text(reply)
     else:
         clean_status = sql.is_enabled(chat.id)
-        if clean_status:
-            clean_status = "Enabled"
-        else:
-            clean_status = "Disabled"
+        clean_status = "Enabled" if clean_status else "Disabled"
         reply = "Bluetext cleaning for <b>{}</b> : <b>{}</b>".format(chat.title, clean_status)
         message.reply_text(reply, parse_mode=ParseMode.HTML)
 
@@ -90,7 +82,7 @@ def add_bluetext_ignore(bot: Bot, update: Update, args: List[str]):
     message = update.effective_message
     chat = update.effective_chat
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
         added = sql.chat_ignore_command(chat.id, val)
         if added:
@@ -98,7 +90,7 @@ def add_bluetext_ignore(bot: Bot, update: Update, args: List[str]):
         else:
             reply = "Command is already ignored."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be ignored."
         message.reply_text(reply)
@@ -111,7 +103,7 @@ def remove_bluetext_ignore(bot: Bot, update: Update, args: List[str]):
     message = update.effective_message
     chat = update.effective_chat
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
         removed = sql.chat_unignore_command(chat.id, val)
         if removed:
@@ -119,7 +111,7 @@ def remove_bluetext_ignore(bot: Bot, update: Update, args: List[str]):
         else:
             reply = "Command isn't ignored currently."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be unignored."
         message.reply_text(reply)
@@ -131,7 +123,7 @@ def add_bluetext_ignore_global(bot: Bot, update: Update, args: List[str]):
 
     message = update.effective_message
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
         added = sql.global_ignore_command(val)
         if added:
@@ -139,7 +131,7 @@ def add_bluetext_ignore_global(bot: Bot, update: Update, args: List[str]):
         else:
             reply = "Command is already ignored."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be ignored."
         message.reply_text(reply)
@@ -151,7 +143,7 @@ def remove_bluetext_ignore_global(bot: Bot, update: Update, args: List[str]):
 
     message = update.effective_message
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
         removed = sql.global_unignore_command(val)
         if removed:
@@ -159,7 +151,7 @@ def remove_bluetext_ignore_global(bot: Bot, update: Update, args: List[str]):
         else:
             reply = "Command isn't ignored currently."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be unignored."
         message.reply_text(reply)
